requirements.txt
cmake>=3.25.0
dlib>=19.22.0
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
customtkinter>=5.2.1
darkdetect>=0.8.0
ttkthemes==3.2.2
tqdm>=4.62.2
pushbullet.py==0.11.0
